+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ﻿# Changelog
 
+## 1.0.3 — 2026-08-08
+
+- Added: Migration dialog to allow users to enable Quality Jobs on existing bills.
+- Fixed: Added tooltips for settings that didn't have one already.
+
+## 1.0.2 — 2026-08-06
+
+- Fixed: Source bill was not properly decremented when items were finished (because the mod creates a separate finisher bill for each item to complete).
+
 ## 1.0.1 — 2026-08-06
 
 - Added: Option to automatically allow the best available colonist to finish jobs (this overrides the skill requirement and makes it follow the best available colonist).